+++ R/anchors.R
@@ -1,7 +1,6 @@
-#' S7 class to define anchors
-#' @param predicates a vector of objects with the S7 class of predicate
-#' @return Object of S7 class anchors
-#' @rdname anchors
+#' @title Defines a new `anchor` object
+#' @param predicates a vector of `predicate` objects
+#' @return A new `anchors` object containing the properties `predicates`
 #' @export
 anchors <- S7::new_class(
   "anchors",