+++ R/predicate.R
@@ -1,8 +1,10 @@
-#' S7 class to define a predicate
+#' @title Defines a predicate object
+#' @description A predicate is built up of a feature column name, a comparison operator and a constant to compare with
+#'  For example a predicate can take the form of x > 10, which captures all the observations that have the feature x greater than 10
 #' @param feature Character. Defines the column that the predicate is applied to
 #' @param operator Binary operator that works with the column and the constant given
 #' @param constant Value to compare `feature` with using `operator`. Can be numeric/character/logical
-#' @rdname anchors
+#' @return A predicate object containing the properties `feature`, `operator`, `constant`
 #' @export
 predicate <- S7::new_class(
   "predicate",