+++ R/zzz.R
@@ -4,6 +4,8 @@ NULL
 
 .onLoad <- function(...) {
   S7::methods_register()
-  envir_to_bounds <<- memoise::memoise(envir_to_bounds)
-  get_reward <<- memoise::memoise(get_reward)
+  if (FALSE) {
+    # WARN: Simply to get tourr in the imports field
+    tourr::grand_tour()
+  }
 }