+++ NEWS.md
@@ -1,5 +1,12 @@
-# kultarr (0.1.0)
+# kultarr (development version)
+
+## Improvements
+
+* Changed underlying algorithm to a breadth-first search algorithm
+
+# kultarr 0.1.0
+
+## Improvements
 
 * Added functions to create anchors
-
 * Added functions to visualize anchors using tourr package